+++ agrest-jaxrs2-openapi/src/test/java/io/agrest/jaxrs2/openapi/Entity_RelationshipsTest.java
@@ -1,23 +1,23 @@
-package io.agrest.jaxrs3.openapi;
+package io.agrest.jaxrs2.openapi;
 
 import io.agrest.DataResponse;
 import io.agrest.annotation.AgRelationship;
-import io.agrest.jaxrs3.openapi.junit.TestOpenAPIBuilder;
+import io.agrest.jaxrs2.openapi.junit.TestOpenAPIBuilder;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.media.ArraySchema;
 import io.swagger.v3.oas.models.media.Schema;
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.core.Configuration;
-import jakarta.ws.rs.core.Context;
 import org.junit.jupiter.api.Test;
 
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.core.Configuration;
+import javax.ws.rs.core.Context;
 import java.util.List;
 import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public class Schema_EntityRelationships_Test {
+public class Entity_RelationshipsTest {
 
     static final OpenAPI oapi = new TestOpenAPIBuilder()
             .addPackage(P5.class)